requirements.txt
opencv-python
numpy
Flask
tqdm
requests
python-dotenv